Implement lazy runtime
In a sense everything is sound without this, as no definitions have any kind of reference
semantics - everything is a function call or indirected, so erroneous insufficiently
lazy mutual reference in initialisation would result in infinite recursion.

The lazy runtime (via the PureScript laziness transform) is provided to ensure the
behaviour is instead to throw, matching the specified semantics and giving a better experience
in the process.